New `--open=vscode` mechanism
* respect existing `.vscode` directory.
* Do not need `README.md`

vscode-rdbg v0.0.9 or later is required to use this feature.

+++ lib/debug/server_dap.rb
@@ -10,44 +10,44 @@ module DEBUGGER__
     SHOW_PROTOCOL = ENV['DEBUG_DAP_SHOW_PROTOCOL'] == '1' || ENV['RUBY_DEBUG_DAP_SHOW_PROTOCOL'] == '1'
 
     def self.setup debug_port
-      dir = Dir.mktmpdir("ruby-debug-vscode-")
-      at_exit{
-        CONFIG[:skip_path] = [//] # skip all
-        FileUtils.rm_rf dir
-      }
-      Dir.chdir(dir) do
-        Dir.mkdir('.vscode')
-        open('README.rb', 'w'){|f|
-          f.puts <<~MSG
-          # Wait for starting the attaching to the Ruby process
-          # This file will be removed at the end of the debuggee process.
-          #
-          # Note that vscode-rdbg extension is needed. Please install if you don't have.
-          MSG
-        }
-        open('.vscode/launch.json', 'w'){|f|
-          f.puts JSON.pretty_generate({
-            version: '0.2.0',
-            configurations: [
-            {
-              type: "rdbg",
-              name: "Attach with rdbg",
-              request: "attach",
-              rdbgPath: File.expand_path('../../exe/rdbg', __dir__),
-              debugPort: debug_port,
-              autoAttach: true,
-            }
-            ]
-          })
-        }
+      if File.directory? '.vscode'
+        dir = Dir.pwd
+      else
+        dir = Dir.mktmpdir("ruby-debug-vscode-")
+        tempdir = true
       end
 
-      cmds = ['code', "#{dir}/", "#{dir}/README.rb"]
+      at_exit do
+        CONFIG[:skip_path] = [//] # skip all
+        FileUtils.rm_rf dir if tempdir
+      end
+
+      key = rand.to_s
+
+      Dir.chdir(dir) do
+        Dir.mkdir('.vscode') if tempdir
+
+        # vscode-rdbg 0.0.9 or later is needed
+        open('.vscode/rdbg_autoattach.json', 'w') do |f|
+          f.puts JSON.pretty_generate({
+            type: "rdbg",
+            name: "Attach with rdbg",
+            request: "attach",
+            rdbgPath: File.expand_path('../../exe/rdbg', __dir__),
+            debugPort: debug_port,
+            localfs: true,
+            autoAttach: key,
+          })
+        end
+      end
+
+      cmds = ['code', "#{dir}/"]
       cmdline = cmds.join(' ')
-      ssh_cmdline = "code --remote ssh-remote+[SSH hostname] #{dir}/ #{dir}/README.rb"
+      ssh_cmdline = "code --remote ssh-remote+[SSH hostname] #{dir}/"
 
       STDERR.puts "Launching: #{cmdline}"
       env = ENV.delete_if{|k, h| /RUBY/ =~ k}.to_h
+      env['RUBY_DEBUG_AUTOATTACH'] = key
 
       unless system(env, *cmds)
         DEBUGGER__.warn <<~MESSAGE
